# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/orders": {
      "post": {
        "operationId": "OrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      },
      "get": {
        "operationId": "OrdersController_list",
        "parameters": [
          {
            "name": "role",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/api/orders/authenticator-search": {
      "get": {
        "operationId": "OrdersController_authenticatorSearch",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/api/orders/{id}": {
      "get": {
        "operationId": "OrdersController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/confirm-review": {
      "patch": {
        "operationId": "OrdersController_confirmReview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/pay": {
      "patch": {
        "operationId": "OrdersController_pay",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/ship-to-authenticator": {
      "patch": {
        "operationId": "OrdersController_shipToAuthenticator",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/ship-to-buyer-direct": {
      "patch": {
        "operationId": "OrdersController_shipToBuyerDirect",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/mark-received": {
      "patch": {
        "operationId": "OrdersController_markReceived",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhotosDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/verdict": {
      "patch": {
        "operationId": "OrdersController_verdict",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerdictDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/evidence": {
      "post": {
        "operationId": "OrdersController_addEvidence",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddEvidenceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      },
      "get": {
        "operationId": "OrdersController_listEvidence",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/dispute-ship": {
      "post": {
        "operationId": "OrdersController_disputeShip",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeReasonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/handover-token": {
      "get": {
        "operationId": "OrdersController_issueHandoverToken",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/qr/scan": {
      "post": {
        "operationId": "OrdersController_scanQr",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QrTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/qr/confirm": {
      "post": {
        "operationId": "OrdersController_confirmQr",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QrConfirmDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/ship-to-buyer": {
      "patch": {
        "operationId": "OrdersController_shipToBuyer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/api/orders/{id}/start-meetup-auth": {
      "patch": {
        "operationId": "OrdersController_startMeetupAuth",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/custody-phone-fallback": {
      "patch": {
        "operationId": "OrdersController_custodyPhoneFallback",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustodyPhoneFallbackDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/complete-meetup": {
      "patch": {
        "operationId": "OrdersController_completeMeetup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/start-meetup-handover": {
      "patch": {
        "operationId": "OrdersController_startMeetupHandover",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/auth-receive-ack": {
      "patch": {
        "operationId": "OrdersController_authReceiveAck",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhotosDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/seller-handover-ack": {
      "patch": {
        "operationId": "OrdersController_sellerHandoverAck",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/submit-verdict-meetup": {
      "patch": {
        "operationId": "OrdersController_submitVerdictMeetup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerdictDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/buyer-receive-ack": {
      "patch": {
        "operationId": "OrdersController_buyerReceiveAck",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/upload-return-photos": {
      "patch": {
        "operationId": "OrdersController_uploadReturnPhotos",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhotosDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/seller-return-ack": {
      "patch": {
        "operationId": "OrdersController_sellerReturnAck",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/dispute": {
      "patch": {
        "operationId": "OrdersController_dispute",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/request-rephoto": {
      "patch": {
        "operationId": "OrdersController_requestRePhoto",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RePhotoRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/cancel-handover": {
      "patch": {
        "operationId": "OrdersController_cancelHandover",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SoftReasonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/complete": {
      "patch": {
        "operationId": "OrdersController_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/orders/{id}/review": {
      "post": {
        "operationId": "OrdersController_review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/api/conversations/order/{orderId}": {
      "get": {
        "operationId": "MessagesController_getOrderMessages",
        "parameters": [
          {
            "name": "orderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Messages"
        ]
      }
    },
...
    "/api/admin/orders/{id}": {
      "get": {
        "operationId": "AdminController_adminOrderDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Certifine API",
    "description": "C2C marketplace with per-category third-party authentication. Generated from the running app — do not hand-edit, and do not hand-edit the client generated from it.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "listingId": {
            "type": "string"
          },
          "authenticatorId": {
            "type": "string"
          },
          "deliveryMethod": {
            "type": "string",
            "enum": [
              "SHIP",
              "MEETUP_AUTH",
              "MEETUP_3WAY",
              "MEETUP_DIRECT"
            ]
          },
          "paymentMethod": {
            "type": "string",
            "enum": [
              "ONLINE_ESCROW",
              "OFFLINE_CASH"
            ]
          },
          "meetupLocation": {
            "type": "string"
          },
          "meetupBranchId": {
            "type": "string"
          },
          "meetupFreeText": {
            "type": "string"
          },
          "offerId": {
            "type": "string"
          }
        },
        "required": [
          "listingId",
          "deliveryMethod",
          "paymentMethod"
        ]
      },
      "TrackingDto": {
        "type": "object",
        "properties": {
          "trackingNo": {
            "type": "string"
          }
        }
      },
      "PhotosDto": {
        "type": "object",
        "properties": {
          "photos": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "photos"
        ]
      },
      "VerdictDto": {
        "type": "object",
        "properties": {
          "verdict": {
            "type": "string",
            "enum": [
              "PASSED",
              "FAILED",
              "INCONCLUSIVE"
            ]
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "verdict"
        ]
      },
      "AddEvidenceDto": {
        "type": "object",
        "properties": {
          "mediaUrl": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "number"
          },
          "kind": {
            "type": "string",
            "enum": [
              "VIDEO",
              "IMAGE"
            ]
          }
        },
        "required": [
          "mediaUrl",
          "mimeType",
          "sizeBytes",
          "kind"
        ]
      },
      "DisputeReasonDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "reason"
        ]
      },
      "QrTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ]
      },
      "QrConfirmDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "photos": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "token"
        ]
      },
      "CustodyPhoneFallbackDto": {
        "type": "object",
        "properties": {
          "sellerPhone": {
            "type": "string"
          },
          "photos": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "sellerPhone",
          "photos"
        ]
      },
      "DisputeDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "reason"
        ]
      },
      "RePhotoRequestDto": {
        "type": "object",
        "properties": {
          "presets": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "comment": {
            "type": "string"
          }
        }
      },
      "SoftReasonDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
      "ReviewDto": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "number"
          },
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "rating"
        ]
      },
...
    }
  }
}